backend/requirements.txt
requests
beautifulsoup4
python-dotenv
fastapi
uvicorn
psycopg2-binary
